add etudiant account and activation key send email verification account verification login and security check //TO DO/// edit and modify data

+++ src/main/java/com/sos/fso/cdoc/insc/entities/Etudiant.java
@@ -65,7 +65,7 @@ public class Etudiant implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "cne")
-    private BigInteger cne;
+    private long cne;
     @Column(name = "optimistic_lock")
     private Integer optimisticLock;
     @Basic(optional = false)
@@ -118,7 +118,7 @@ public class Etudiant implements Serializable {
         this.idEtudiant = idEtudiant;
     }
 
-    public Etudiant(Integer idEtudiant, String cin, BigInteger cne, String nom, String prenom, Date dateNaissance, String email) {
+    public Etudiant(Integer idEtudiant, String cin, long cne, String nom, String prenom, Date dateNaissance, String email) {
         this.idEtudiant = idEtudiant;
         this.cin = cin;
         this.cne = cne;
@@ -144,11 +144,11 @@ public class Etudiant implements Serializable {
         this.cin = cin;
     }
 
-    public BigInteger getCne() {
+    public long getCne() {
         return cne;
     }
 
-    public void setCne(BigInteger cne) {
+    public void setCne(long cne) {
         this.cne = cne;
     }
